+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ### Patch Changes
 
+- Updated dependencies [80d4dde]
+  - ai@7.0.0-beta.23
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@7.0.0-beta.22
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [34fd051]
   - ai@7.0.0-beta.21